Insert the "spark" prefix into SparkUI responses.
Hackishly insert the spark link prefix ("/spark/") into docs
returned from the Spark UI. This makes the resulting page work
as expected, including links, images, and css/js.

It may be too fragile, so if there is a better way to do this, I
am open to changing the approach.